geostats_server/requirements.txt
Flask
SQLAlchemy
psycopg2-binary
python-geohash
jsonify
